requirements.txt
#torch==2.7.1
#torchvision>=0.24.1
opencv-python>=4.9.0.80
diffusers>=0.31.0
transformers==4.49.0
tokenizers>=0.20.3
accelerate>=1.1.1
tqdm
imageio
easydict
ftfy
dashscope
imageio-ffmpeg
#flash_attn
gradio>=5.0.0
numpy
boto3
decord
wandb[media]
omegaconf
einops
av==13.1.0
git+https://github.com/openai/CLIP.git
open_clip_torch
starlette
lmdb
matplotlib
sentencepiece
pydantic
scikit-image
ffmpeg-python
ffprobe-python
outlines
torchinfo
